refact(api): migrate local BaseModel subclasses — analytics + knowledge domains (#6042)
Phase 1 of #6042: move response/request schemas out of endpoint files into domain schema files.

analytics_precommit.py (11 classes + 2 enums) → schemas_analytics.py
  - CheckSeverity, CheckCategory (enums)
  - CheckResult, CommitCheckResult, HookConfig, CheckDefinition, HookStatus
  - CheckToggleResponse, HookConfigUpdateResponse, HookInstallResponse
  - CommonIssueItem, PrecommitSummaryResponse
  - CategoryItem → renamed PrecommitCategoryItem (avoid future collisions)

error_monitoring.py (8 classes) → schemas_analytics.py
  - ErrorMonitoringDataResponse, ErrorMonitoringClearResponse
  - ErrorMonitoringTestErrorResponse, ErrorMonitoringResolveResponse
  - ErrorMonitoringAlertThresholdResponse, ErrorMonitoringCleanupResponse
  - TestErrorRequest, AlertThresholdRequest
  - Also removes pre-existing duplicate import of error_boundaries

knowledge_tags.py (13 classes) → schemas_knowledge.py
  - KnowledgeTagFactTagsResponse, KnowledgeTagGetFactTagsResponse
  - KnowledgeTagSearchResponse, KnowledgeTagListResponse
  - KnowledgeTagBulkResponse, KnowledgeTagRenameResponse
  - KnowledgeTagDeleteResponse, KnowledgeTagMergeResponse
  - KnowledgeTagFactsByTagResponse, KnowledgeTagInfoResponse
  - KnowledgeTagStyleInfo, KnowledgeTagStyleResponse
  - KnowledgeTagDeleteStyleResponse
