+++ autobot-backend/api/knowledge_tags.py
@@ -28,16 +28,27 @@ Related Issues: #77 (Tags), #185 (Split), #209 (Knowledge split), #409 (Tag CRUD
 
 import logging
 import re
-from typing import Any, Dict, List, Optional
+from typing import Optional
 
 from fastapi import APIRouter, Depends, HTTPException, Path, Query
-from pydantic import BaseModel
 from starlette.requests import Request
 
 from api.schemas_knowledge import (
     AddTagsRequest,
     BulkTagRequest,
     FactIdValidator,
+    KnowledgeTagBulkResponse,
+    KnowledgeTagDeleteResponse,
+    KnowledgeTagDeleteStyleResponse,
+    KnowledgeTagFactsByTagResponse,
+    KnowledgeTagFactTagsResponse,
+    KnowledgeTagGetFactTagsResponse,
+    KnowledgeTagInfoResponse,
+    KnowledgeTagListResponse,
+    KnowledgeTagMergeResponse,
+    KnowledgeTagRenameResponse,
+    KnowledgeTagSearchResponse,
+    KnowledgeTagStyleResponse,
     MergeTagsRequest,
     RemoveTagsRequest,
     RenameTagRequest,
@@ -50,134 +61,6 @@ from constants.threshold_constants import QueryDefaults
 from knowledge_factory import get_or_create_knowledge_base
 
 logger = logging.getLogger(__name__)
-
-
-# ---------------------------------------------------------------------------
-# Response models for knowledge_tags endpoints
-# ---------------------------------------------------------------------------
-
-
-class KnowledgeTagFactTagsResponse(BaseModel):
-    """Response for POST/DELETE/GET /fact/{fact_id}/tags."""
-
-    status: str
-    fact_id: str
-    tags: List[Any]
-    message: Optional[str] = None
-    added_count: Optional[int] = None
-    removed_count: Optional[int] = None
-
-
-class KnowledgeTagGetFactTagsResponse(BaseModel):
-    """Response for GET /fact/{fact_id}/tags."""
-
-    status: str
-    fact_id: str
-    tags: List[Any]
-
-
-class KnowledgeTagSearchResponse(BaseModel):
-    """Response for POST /tags/search."""
-
-    status: str
-    facts: List[Any]
-    total_count: int
-    tags_searched: List[str]
-    match_all: bool
-    limit: int
-    offset: int
-
-
-class KnowledgeTagListResponse(BaseModel):
-    """Response for GET /tags."""
-
-    status: str
-    tags: List[Any]
-    total_count: int
-
-
-class KnowledgeTagBulkResponse(BaseModel):
-    """Response for POST /tags/bulk."""
-
-    status: str
-    operation: str
-    processed_count: int
-    failed_count: int
-    results: List[Any]
-
-
-class KnowledgeTagRenameResponse(BaseModel):
-    """Response for PUT /tags/{tag_name}."""
-
-    status: str
-    old_tag: str
-    new_tag: str
-    affected_count: int
-    message: str
-
-
-class KnowledgeTagDeleteResponse(BaseModel):
-    """Response for DELETE /tags/{tag_name}."""
-
-    status: str
-    tag: str
-    affected_count: int
-    message: str
-
-
-class KnowledgeTagMergeResponse(BaseModel):
-    """Response for POST /tags/merge."""
-
-    status: str
-    source_tags: List[str]
-    target_tag: str
-    affected_count: int
-    message: str
-
-
-class KnowledgeTagFactsByTagResponse(BaseModel):
-    """Response for GET /tags/{tag_name}/facts."""
-
-    status: str
-    tag: str
-    facts: List[Any]
-    total_count: int
-    returned_count: int
-    limit: int
-    offset: int
-    has_more: bool
-
-
-class KnowledgeTagInfoResponse(BaseModel):
-    """Response for GET /tags/{tag_name}/info."""
-
-    status: str
-    tag: str
-    fact_count: int
-
-
-class KnowledgeTagStyleInfo(BaseModel):
-    color: Optional[str] = None
-    icon: Optional[str] = None
-    description: Optional[str] = None
-    is_default: bool
-
-
-class KnowledgeTagStyleResponse(BaseModel):
-    """Response for PATCH/GET /tags/{tag_name}/style."""
-
-    status: str
-    tag: str
-    style: Dict[str, Any]
-    message: Optional[str] = None
-
-
-class KnowledgeTagDeleteStyleResponse(BaseModel):
-    """Response for DELETE /tags/{tag_name}/style."""
-
-    status: str
-    tag: str
-    message: str
 
 
 # Issue #380: Pre-compiled regex for tag prefix validation
